restructure the code: Introduce new Classes: - glimpseNetwork.py is responsible for training variables whose does extract information from images - location network - training bandwidth for locations - glimpseSensor - should produce multiple patches for a glimpse. Work in Progress.
