+++ locationNetwork.py
